added kafka commands
