+++ kafka/consume/message/src/pkg/kafkautil/kafkautil.go
