+++ kafka/list/topic/src/go.mod
@@ -1,0 +1,9 @@
+module github.com/ZupIT/ritchie-formulas/kafka/src
+
+go 1.13
+
+require (
+	github.com/Shopify/sarama v1.26.1
+	github.com/hashicorp/go-uuid v1.0.2
+	github.com/manifoldco/promptui v0.7.0
+)
